+++ upload/garrysmod/gamemodes/clockwork/plugins/trackstats/plugin/cl_hooks.lua
@@ -1,7 +1,0 @@
---[[
-© CloudSixteen.com do not share, re-distribute or modify
-without permission of its author (kurozael@gmail.com).
-
-Clockwork was created by Conna Wiles (also known as kurozael.)
-https://creativecommons.org/licenses/by-nc-nd/3.0/legalcode
---]]